Move privateuse1 test out of test_utils and make them serial
